+++ backend/UniGuesser.Application/UseCases/Places/DeletePlace/DeletePlaceHandler.cs
@@ -1,31 +1,24 @@
 ﻿using MediatR;
-using UniGuesser.Application.Services.SaveFileService;
-using UniGuesser.Domain.Entities;
+using UniGuesser.Application.Services;
 using UniGuesser.Domain.Exceptions;
 using UniGuesser.Domain.Repositories;
 using UniGuesser.Domain.ValueObjects.Enumerations;
 
-namespace UniGuesser.Application.UseCases.Places.DeletePlace
+namespace UniGuesser.Application.UseCases.Places.DeletePlace;
+
+public class DeletePlaceHandler(IPlacesRepository placesRepository, IFileService fileService)
+    : IRequestHandler<DeletePlaceCommand, Unit>
 {
-    public class DeletePlaceHandler(IPlacesRepository placesRepository, IFileService fileService) : IRequestHandler<DeletePlaceCommand, Unit>
+    public async Task<Unit> Handle(DeletePlaceCommand request, CancellationToken cancellationToken)
     {
-        public async Task<Unit> Handle(DeletePlaceCommand request, CancellationToken cancellationToken)
-        {
-            Place? place = await placesRepository.GetAsync(request.guid);
+        var place = await placesRepository.GetAsync(request.guid);
 
-            if (place is null)
-            {
-                throw new PlacesExceptions.PlaceNotFoundException(request.guid);
-            }
+        if (place is null) throw new PlacesExceptions.PlaceNotFoundException(request.guid);
 
-            if (place.ImageType == ImageType.File)
-            {
-                fileService.DeleteFile(place.ImageUrl);
-            }
+        if (place.ImageType == ImageType.File) fileService.DeleteFile(place.ImageUrl);
 
-            var deleteResult = await placesRepository.DeleteAsync(place.Id);
+        var deleteResult = await placesRepository.DeleteAsync(place.Id);
 
-            return Unit.Value;
-        }
+        return Unit.Value;
     }
 }
